mwc_gallery/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

--extra-index-url=http://dist.gondor.io/dev/
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/fresh-start/

-e git+http://github.com/jlecker/django-rubberstamp.git@master#egg=rubberstamp
-e git+http://github.com/MidwestCommunications/mimesis.git@master#egg=mimesis

django-photologue==2.2
django-taggit==0.8.0
nashvegas==0.6.3
django_compressor==0.9.2
psycopg2==2.3
# See http://pip-installer.org/requirement-format.html for more information.
